Translate legacy go flag args to Cobra args
This commit intercepts the command line args passed to `faas-cli` and
attempts to translate them from the deprecated go flag based syntax
(`faas-cli -action xxx`) to the new Cobra verb/noun based syntax
(`faas-cli xxx`), it also translates a frozen set of legacy flags (with
the go-style single-dash) into a GNU style double-dash.

Note that some special cases are included:
- changing the delete action to remove
- passing the function name as a noun to remove rather than as an arg to
`-name`
- it also handles the legacy format where args are passed after =
(`-name=fnname`).

If the translation results in a new set of args then a message is
displayed to the user (stderr) telling warning that they are using the
deprecated cli syntax and also prints the new syntax command that is
being executed and which they should use going forward.

Any errors thrown during translation result in the command failing with
it printing the error cause to stderr.

Signed-off-by: John McCabe <john@johnmccabe.net>