Use snake rather than camel case for go src
This renames the `fetchTemplates.go` file to use snake case. The
convention appears to be for snakecase - as observed in both the Go and
Kubernetes source. For example heres a random selection of source files.

-
https://github.com/kubernetes/kubernetes/blob/master/pkg/kubeapiserver/default_storage_factory_builder.go
-
https://github.com/kubernetes/kubernetes/blob/master/pkg/kubectl/bash_comp_utils.go
-
https://github.com/golang/go/blob/master/src/compress/bzip2/move_to_front.go

Note that the language spec does not set a hard rule for source file
names, only for package names, but making this change for consistency.

Signed-off-by: John McCabe <john@johnmccabe.net>